refactor get mesaures
