+++ helpers/measures.js
@@ -8,6 +8,12 @@ const { buildDateString } = require('helpers/utils');
 const moment = require('moment');
 const validation = require('helpers/validation');
 const parse = require('helpers/parse');
+const filterMetricsHelper = require('helpers/filterMetrics');
+const rayg = require('helpers/rayg');
+const Entity = require('models/entity');
+const groupBy = require('lodash/groupBy');
+const get = require('lodash/get');
+
 
 const applyLabelToEntities = (entities) => {
   entities.forEach(entity => {
@@ -132,12 +138,102 @@ const validateEntities = async (entities) => {
   }
 }
 
+const getMeasureEntities = async({ measureCategory, themeCategory, where, user }) => {
+  where = { categoryId: measureCategory.id, ...where };
+
+  let measureEntities = await Entity.findAll({
+    where,
+    include: [{
+      separate: true,
+      model: EntityFieldEntry,
+      include: CategoryField
+    },{
+      // get direct parent of measure i.e. outcome statement
+      model: Entity,
+      as: 'parents',
+      include: [{
+        model: Category
+      }, {
+        // get direct parents of outcome statement i.e. theme ( and possibly another outcome statement )
+        model: Entity,
+        as: 'parents',
+        include: [{
+          separate: true,
+          model: EntityFieldEntry,
+          include: CategoryField
+        }, {
+          model: Category
+        }]
+      }]
+    }]
+  });
+
+  if (user) {
+    measureEntities = await filterMetricsHelper.filterMetrics(user, measureEntities);
+  }
+
+  return measureEntities.map(entity => {
+    const theme = get(entity, 'parents[0].parents').find(parentEntity => {
+      return parentEntity.categoryId === themeCategory.id;
+    });
+
+    const themeName = theme.entityFieldEntries.find(fieldEntry => {
+      return fieldEntry.categoryField.name === 'name';
+    });
+
+    const entityMapped = {
+      id: entity.id,
+      publicId: entity.publicId,
+      theme: themeName.value
+    };
+
+    entity.entityFieldEntries.map(entityfieldEntry => {
+      entityMapped[entityfieldEntry.categoryField.name] = entityfieldEntry.value;
+    });
+
+    return entityMapped;
+  });
+}
+
+const groupMeasures = (measures) => {
+  const measureEntitiesGrouped = groupBy(measures, measure => {
+    return measure.groupID;
+  });
+
+  const measureGroups = Object.values(measureEntitiesGrouped).reduce((measureGroups, group) => {
+    const groupMeasure = group.find(measure => measure.filter === 'RAYG');
+    // if no RAYG row then ignore as its not shown on dashboard
+    if(!groupMeasure) {
+      return measureGroups;
+    }
+
+    const nonRaygRows = group.filter(measure => measure.filter !== 'RAYG');
+
+    const measuresGroupedByMetricId = groupBy(nonRaygRows, entity => entity.metricID);
+    groupMeasure.children = Object.values(measuresGroupedByMetricId).map(measures => {
+      const measuresSortedByDate = measures.sort((a, b) => moment(b.date, 'DD/MM/YYYY').valueOf() - moment(a.date, 'DD/MM/YYYY').valueOf());
+      measuresSortedByDate[0].colour = rayg.getRaygColour(measuresSortedByDate[0]);
+      return measuresSortedByDate[0];
+    });
+
+    groupMeasure.colour = rayg.getRaygColour(groupMeasure);
+
+    measureGroups.push(groupMeasure);
+
+    return measureGroups;
+  }, []);
+
+  return measureGroups;
+} 
+
 module.exports = {
   applyLabelToEntities,
   calculateUiInputs,
   getEntityFields,
   getCategory,
   getMeasureEntitiesFromGroup,
+  getMeasureEntities,
+  groupMeasures,
   removeBlankEntityInputValues,
   validateFormData,
   validateEntities
